Bug fix for liveness probe.
Turns out when really busy (e.g. someone is uploading data) the api can
take more than the default timeout of 1 second to reply to a liveness
probe... which caused k8s to restart the api pod. Not good.

This increases that timeout and only does a liveness probe every 5 min