Refactor redis cache (#430)
Refactors the entityservice.cache to make it easier to understand and extend. Also attempts to clean up import code to be more explicit about where functions from the cache module.

One of the few (minor) changes is at the start of a run the current number of comparisons is set to zero, instead of waiting for the first chunk of similarities to be completed.

The logging config now includes separate logging levels for the cache.